+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+-   Add initial tenants listing page
+
 ## [0.10.2] - 2024-01-26
 
 -   Fix typo on user roles section on user details page.